Split out imageSizeFromAttachments
